+++ packages/backend/scripts/compare-flat/executeFindSimilar.ts
@@ -5,40 +5,26 @@ import { powerdiff } from '../powerdiffCore'
 import { FindSimilarCommand } from './cli'
 import {
   ALL_CONFIGS,
-  colorMap,
   computeComparisonBetweenProjects,
   computeStackSimilarity,
   decodeProjectPath,
-  formatHeader,
   getMostSimilar,
   needsToBe,
-  printComparisonBetweenProjects,
 } from './common'
+import {
+  colorMap,
+  formatHeader,
+  printComparisonBetweenProjects,
+} from './output'
 
 export async function executeFindSimilar(
   command: FindSimilarCommand,
 ): Promise<void> {
   const { name, chain } = decodeProjectPath(command.projectPath)
-
-  const projects = ALL_CONFIGS.filter((p) => p.id.toString() === name)
-  needsToBe(
-    projects.length <= 1,
-    `More than one project found matching ${
-      command.projectPath
-    } ${projects.toString()}`,
-  )
-  needsToBe(
-    projects.length === 1,
-    `No project found matching ${command.projectPath}`,
-  )
-  const project = projects[0]
-  needsToBe(
-    project.display.provider !== undefined,
-    `Project ${project.display.name} has no provider`,
-  )
+  const project = getProject(name, command)
 
   const { matrix: perProjectMatrix } = await computeStackSimilarity(
-    project.display.provider,
+    project.stack,
   )
   const mostSimilar = getMostSimilar(perProjectMatrix)
 
@@ -66,5 +52,28 @@ export async function executeFindSimilar(
     const path1 = path.join('discovery', name, chain, '.flat')
     const path2 = path.join('discovery', otherName, otherChain, '.flat')
     powerdiff(path1, path2)
+  }
+}
+
+function getProject(name: string, command: FindSimilarCommand) {
+  const projects = ALL_CONFIGS.filter((p) => p.id.toString() === name)
+  needsToBe(
+    projects.length <= 1,
+    `More than one project found matching ${
+      command.projectPath
+    } ${projects.toString()}`,
+  )
+  needsToBe(
+    projects.length === 1,
+    `No project found matching ${command.projectPath}`,
+  )
+  const project = projects[0]
+  needsToBe(
+    project.display.provider !== undefined,
+    `Project ${project.display.name} has no provider`,
+  )
+  return {
+    id: project.id,
+    stack: project.display.provider,
   }
 }
